+++ src/test/java/http/RequestHeaderFirstLineTest.java
@@ -1,13 +1,13 @@
-package utils;
+package http;
 
-import http.RequestHeaderFirstLine;
 import org.junit.jupiter.api.Test;
+import utils.RequestHeaderUtils;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class RequestHeaderUtilsTest {
+public class RequestHeaderFirstLineTest {
     @Test
-    void pathParser() {
+    void requestHeaderFirstLineParse() {
         RequestHeaderFirstLine result = RequestHeaderUtils.parse("GET /index.html HTTP/1.1");
         assertThat(result.getPath()).isEqualTo("/index.html");
     }
